fix(flagd): address code review feedback for WASM resolver
- Load WASM from classpath instead of hardcoded path
- Use static SecureRandom for efficiency
- Fix JSON exception handling to not terminate stateWatcher thread
- Implement shutdown() method properly
- Change generic Exception to GeneralError
- Fix ImmutableMetadataDeserializer to preserve value types
- Move OBJECT_MAPPER config to static block
- Make stateWatcher a class field for proper shutdown
- Remove dead code (getFlagMetadata, FLAGD_SELECTOR_KEY)
- Extract serializers/deserializers to jackson package
